PROG-LANG: Complex numbers class implemented

+++ programming-languages/practices/geometry/Point.java
@@ -1,8 +1,8 @@
 public class Point {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -13,15 +13,15 @@ public class Point {
     }
 
     public Point mirror(Point reflection) {
-        int mirroredX = 2 * reflection.x - x;
-        int mirroredY = 2 * reflection.y - y;
+        double mirroredX = 2 * reflection.x - x;
+        double mirroredY = 2 * reflection.y - y;
 
         return new Point(mirroredX, mirroredY);
     }
 
     public double distance(Point p) {
-        int dx = this.x - p.x;
-        int dy = this.y - p.y;
+        double dx = this.x - p.x;
+        double dy = this.y - p.y;
 
         double d = Math.sqrt(dx*dx + dy*dy);
         return d;
